+++ src/main/java/gregtech/client/event/ClientEventHandler.java
@@ -1,6 +1,7 @@
 package gregtech.client.event;
 
 import gregtech.client.renderer.handler.BlockPosHighlightRenderer;
+import gregtech.client.renderer.handler.MultiblockPreviewRenderer;
 import gregtech.client.utils.DepthTextureUtil;
 import gregtech.client.renderer.handler.TerminalARRenderer;
 import gregtech.client.renderer.handler.ToolOverlayRenderer;
@@ -43,6 +44,7 @@ public class ClientEventHandler {
     public static void onRenderWorldLast(RenderWorldLastEvent event) {
         DepthTextureUtil.renderWorld(event);
 //        GTParticleManager.renderWorld(event);
+        MultiblockPreviewRenderer.renderWorldLastEvent(event);
         BlockPosHighlightRenderer.renderWorldLastEvent(event);
         TerminalARRenderer.renderWorldLastEvent(event);
     }
